// src/core/types/employee.d.ts
export declare interface IEmployee {
	id: number;
	fullName: string;
	position: string;
	titleOfCourtesy: string;
	birthDate: Date;
	hireDate: Date;
	address: string;
	city: string;
	region: string;
	postalCode: string;
	country: string;
	homePhone: string;
	extension: string;
	photo: string;
	notes: string;
	reportsTo: number;
}